spotify-analytics-platform/requirements.txt
apache-airflow==2.7.1
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.7
sqlalchemy==2.0.19
streamlit==1.28.0
plotly==5.17.0
python-dotenv==1.0.0
